- Adds :$raw parameter to Clutter::Actor.get-child-at-index - Adds Clutter::BoxLayout.setup - Fixes $proper-name resolution in Clutter::FlowLayout.setup - Adds defaults to Clutter::Point.new which will allow it to be called with no arguments. - Fixes initialization logic and adds setup method to Clutter::ScrollActor - Adds ancestry logic and identity constructor to Clutter::Text - Updates t/12-scroll-actor.t. Current state is now COMPLETE.
+ NOTE TO SELF: Steps will need to be taken across the ENTIRE GTK LINE 
to insure that :$raw parameters are added to any object that returns a 
method. This will give callers the option to cast the returned pointer 
to something else. This will come in handy when it is required to cast a 
returned value to a descendent class to prevent an ambiguous call. 
Especially to hierarchies that implement an identity constructor all the 
way down to the base class!!